Put calc summary file operations in CalcInfo class
